% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05-CWL-class.R
\docType{methods}
\name{asJSON}
\alias{asJSON}
\alias{asJSON,ANY-method}
\title{Convert a object slots/fields to a JSON}
\usage{
asJSON(object, ...)

\S4method{asJSON}{ANY}(object, ...)
}
\arguments{
\item{object}{object, could be S4/R5 object, for example, class CWL, SimpleList.}

\item{...}{other parameters passed to \code{toJSON}}
}
\description{
Convert a object slots/fields to a JSON
}

