% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05-CWL-class.R
\docType{methods}
\name{asList}
\alias{asList}
\alias{asList,ANY-method}
\alias{asList,CWL-method}
\alias{asList,SimpleList-method}
\title{Convert a object slots/fields to a list}
\usage{
asList(object, ...)

\S4method{asList}{ANY}(object, ...)

\S4method{asList}{CWL}(object, ...)

\S4method{asList}{SimpleList}(object, ...)
}
\arguments{
\item{object}{object, could be S4/R5 object, for example, class CWL, SimpleList.}

\item{...}{other parameters passed}
}
\description{
Doesn't like \code{as.list}, only fields and slots are converted,
prepare a object to be conveted to YAML/JSON.
}

