% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/60-CommandLineTool-class.R
\docType{class}
\name{CommandOutputSchema-class}
\alias{CommandOutputSchema}
\alias{CommandOutputSchema-class}
\title{CommandOutputSchema}
\description{
CommandOutputSchema
}
\section{Fields}{

\describe{
\item{\code{outputBinding}}{[CommandOutputBinding] Describes how to handle
the concrete outputs of a process step (such as files created by a
program) and describe them in the process output parameter.}
}}
\examples{
CommandOutputSchema()
}

