% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CWLAllClasses.R
\docType{class}
\name{CreateFileRequirement-class}
\alias{CreateFileRequirement}
\alias{CreateFileRequirement-class}
\title{CreateFileRequirement Class}
\description{
Define a list of files that must be created and placed by the
workflow platform in the designated output directory prior to
executing the command line tool. See FileDef for details.
}
\section{Fields}{

\describe{
\item{\code{fileDef}}{(FileDefList) The list of files.}
}}

