% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/10-Datatype-Class.R
\docType{class}
\name{PrimitiveSingleEnum-class}
\alias{ComplexEnum}
\alias{ComplexSingleEnum-class}
\alias{DatatypeEnum}
\alias{DatatypeSingleEnum-class}
\alias{PrimitiveEnum}
\alias{PrimitiveSingleEnum-class}
\title{Pre-defiend enums}
\description{
Please check \code{cwl:::.CWL.Pritimive}, \code{cwl:::.CWL.Complex}.
}
\examples{
PrimitiveEnum()
PrimitiveEnum("boolean")
ComplexEnum("record")
DatatypeEnum("map")
}

