% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/30-ProcessRequirement-class.R
\docType{class}
\name{EnvironmentDef-class}
\alias{EnvironmentDef}
\alias{EnvironmentDef-class}
\title{EnvironmentDef Class}
\description{
Define an environment variable that will be set in the runtime
environment by the workflow platform when executing the command
line tool. May be the result of executing an expression, such as
getting a parameter from input.
}
\section{Fields}{

\describe{
\item{\code{envName}}{(character) The environment variable name.}

\item{\code{envValue}}{(characterORExpression) The environment variable value.}
}}

