% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/30-ProcessRequirement-class.R
\docType{class}
\name{EnvVarRequirement-class}
\alias{EnvVarRequirement}
\alias{EnvVarRequirement-class}
\title{EnvVarRequirement Class}
\description{
Define a list of environment variables which will be set in the
execution environment of the tool. See EnvironmentDef for details.
}
\section{Fields}{

\describe{
\item{\code{envDef}}{(EnvironmentDefList) The list of environment
variables.}
}}

