% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03-ProcessRequirement-class.R
\docType{class}
\name{ExpressionEngineRequirement-class}
\alias{ExpressionEngineRequirement}
\alias{ExpressionEngineRequirement-class}
\title{ExpressionEngineRequirement Class}
\description{
Define an expression engine, as described in Expressions.
}
\section{Fields}{

\describe{
\item{\code{id}}{(character) Used to identify the expression engine in the
engine field of Expressions.}

\item{\code{requirements}}{[ProcessRequirement]Requirements to run this
expression engine, such as DockerRequirement for specifying a
container with the engine.}

\item{\code{engineCommand}}{[character] The command line to invoke the
expression engine.}

\item{\code{engineConfig}}{[character] Additional configuration or code
fragments that will also be passed to the expression engine. The
semantics of this field are defined by the underlying expression
engine. Intended for uses such as providing function definitions
that will be called from CWL expressions.}
}}

