% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/50-ExpressionTool-class.R
\docType{class}
\name{ExpressionTool-class}
\alias{ExpressionTool}
\alias{ExpressionTool-class}
\title{ExpressionTool Class}
\description{
Execute an expression as a process step.
}
\section{Fields}{

\describe{
\item{\code{expression}}{(Expression) The expression to execute. The
expression must return a JSON object which matches the output
parameters of the ExpressionTool.}
}}
\examples{
ExpressionTool(expression =
                  Expression(engine = "cwl:JsonPointer",
                             script = "$job.inputs['threads']"))
}

