% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CWLAllClasses.R
\docType{class}
\name{File-class}
\alias{File}
\alias{File-class}
\title{File Class}
\value{
File class generator
}
\description{
File Class
}
\section{Fields}{

\describe{
\item{\code{path}}{(character) The path to the file.}

\item{\code{checksum}}{[character] Optional hash code for validating file
integrity. Currently must be in the form "sha1$ + hexidecimal
string" using the SHA-1 algorithm.}

\item{\code{size}}{[numeric] Optional file size.}

\item{\code{secondaryFile}}{[FileList] A list of additional files that are
associated with the primary file and must be transferred alongside
the primary file. Examples include indexes of the primary file, or
external references which must be included when loading primary
document. A file object listed in secondaryFiles may itself include
secondaryFiles for which the same rules apply.}
}}
\examples{
f1 <- new("File")
f2 <- File()
FileList(f1, f2)
}

