% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CWLAllClasses.R
\docType{class}
\name{FileDef-class}
\alias{FileDef}
\alias{FileDef-class}
\title{FileDef Class}
\description{
Define a file that must be placed by in the designated output
directory prior to executing the command line tool. May be the
result of executing an expression, such as building a configuration
file from a template.
}
\section{Fields}{

\describe{
\item{\code{filename}}{(characterORExpression) The name of the file to
create in the output directory.}

\item{\code{fileContent}}{(characterORExpression) If the value is a string
literal or an expression which evalutes to a string, a new file
must be created with the string as the file contents. If the value
is an expression that evaluates to a File object, this indicates
the referenced file should be added to the designated output
directory prior to executing the tool. Files added in this way may
be read-only, and may be implemented through bind mounts or file
system links in such a way as to avoid unecessary copying of the
input file.}
}}

