% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/30-ProcessRequirement-class.R
\name{FileDefList}
\alias{FileDefList}
\alias{FileDefList-class}
\title{FileDefList}
\usage{
FileDefList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\description{
FileDefList
}

