% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/10-Datatype-Class.R
\name{FileList}
\alias{FileList}
\alias{FileList-class}
\title{FileList Class}
\usage{
FileList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\description{
FileList Class
}

