% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04-Process-class.R
\docType{class}
\name{InputParameter-class}
\alias{InputParameter}
\alias{InputParameter-class}
\title{InputParameter Class}
\description{
InputParameter Class
}
\section{Fields}{

\describe{
\item{\code{id}}{(character) The unique identifier for this parameter object.}

\item{\code{inputBinding}}{[Binding] Describes how to handle the inputs of
a process and convert them into a concrete form for execution, such
as command line parameters.}
}}

