% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CWLAllClasses.R
\name{InputParameterList}
\alias{InputParameterList}
\alias{InputParameterList-class}
\title{InputParameterList}
\usage{
InputParameterList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\description{
InputParameterList
}

