% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CWLAllClasses.R
\docType{class}
\name{InputSchema-class}
\alias{InputSchema}
\alias{InputSchema-class}
\title{InputSchema Class}
\description{
InputSchema Class
}
\section{Fields}{

\describe{
\item{\code{inputBinding}}{[Binding] Describes how to handle a value in
the input object convert it into a concrete form for execution,
such as command line parameters.}
}}

