% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CWLAllClasses.R
\docType{class}
\name{OutputParameter-class}
\alias{OutputParameter}
\alias{OutputParameter-class}
\title{OutputParameter Class}
\description{
OutputParameter Class
}
\section{Fields}{

\describe{
\item{\code{id}}{(character) The unique identifier for this parameter object.}
}}

