% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/40-Process-class.R
\name{OutputParameterList}
\alias{OutputParameterList}
\alias{OutputParameterList-class}
\title{OutputParameterList}
\usage{
OutputParameterList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\description{
OutputParameterList
}

