% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CWLAllClasses.R
\docType{class}
\name{Parameter-class}
\alias{Parameter}
\alias{Parameter-class}
\title{Paramter class (reference class)}
\description{
Define an input or output parameter to a process.
}
\section{Fields}{

\describe{
\item{\code{type}}{[ANY] Specify valid types of data that may be assigned
to this parameter.}

\item{\code{label}}{[character] A short, human-readable label of this
parameter object.}

\item{\code{description}}{[character] A long, human-readable description
of this parameter object.}

\item{\code{streamable}}{[logical] Currently only applies if type is
File. A value of true indicates that the file is read or written
sequentially without seeking. An implementation may use this flag
to indicate whether it is valid to stream file contents using a
named pipe. Default: false.}

\item{\code{default}}{[ANY] The default value for this parameter if not
provided in the input object.}
}}
\examples{
Parameter(type = "integer", label = "thread",
         description = "Specify the thread #",
         default = 0)
}

