% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/30-ProcessRequirement-class.R
\docType{class}
\name{ProcessRequirement-class}
\alias{ProcessRequirement}
\alias{ProcessRequirement-class}
\title{ProcessRequiremen Class}
\description{
A process requirement declares a prerequisite that may or must be
fufilled before executing a process. See Process.hints and
Process.requirements. Process requirements are the primary
mechanism for specifying extensions to the CWL core specification.
}
\section{Fields}{

\describe{
\item{\code{(character)}}{The specific requirement type.}
}}

