% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/30-ProcessRequirement-class.R
\docType{class}
\name{ScatterFeatureRequirement-class}
\alias{ScatterFeatureRequirement}
\alias{ScatterFeatureRequirement-class}
\title{ScatterFeatureRequirement Class}
\description{
Indicates that the workflow platform must support the scatter and
scatterMethod fields of (WorkflowStep)(#workflowstep).
}

