% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CWLAllClasses.R
\docType{class}
\name{Schema-class}
\alias{Schema}
\alias{Schema-class}
\title{Schema Class}
\description{
A schema defines a parameter type.
}
\section{Fields}{

\describe{
\item{\code{type}}{(ANY) The data type of this parameter.}

\item{\code{fileds}}{[SchemaList] When type is record, defines the fields of the
record.}

\item{\code{symbols}}{[character] When type is enum, defines the set of valid symbols.}

\item{\code{items}}{[ANY] When type is array, defines the type of the array
elements.}

\item{\code{values}}{[ANY] When type is map, defines the value type for the
key/value pairs.}
}}

