% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/40-Process-class.R
\name{SchemaDefList}
\alias{SchemaDefList}
\alias{SchemaDefList-class}
\title{SchemaDefList}
\usage{
SchemaDefList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\description{
SchemaDefList
}

