% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/40-Process-class.R
\name{SchemaList}
\alias{SchemaList}
\alias{SchemaList-class}
\title{SchemaList}
\usage{
SchemaList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\description{
SchemaList
}

