% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CWLAllClasses.R
\name{setListClass}
\alias{setListClass}
\title{List Class generator.}
\usage{
setListClass(elementType = NULL, suffix = "List", contains = NULL)
}
\arguments{
\item{elementType}{[character]}

\item{suffix}{[character] default is "List"}

\item{contains}{[character] class name.}
}
\value{
S4 class generator
}
\description{
Take simple approach to extens list class, please note that in IRanges package
there is more generic List class with utilities.
}

