% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/30-ProcessRequirement-class.R
\docType{class}
\name{SubworkflowFeatureRequirement-class}
\alias{SubworkflowFeatureRequirement}
\alias{SubworkflowFeatureRequirement-class}
\title{SubworkflowFeatureRequirement Class}
\description{
Indicates that the workflow platform must support nested workflows
in the run field of (WorkflowStep)(#workflowstep).
}

