% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/07-Workflow-class.R
\docType{class}
\name{Workflow-class}
\alias{Workflow}
\alias{Workflow-class}
\title{Workflow}
\description{
A workflow is a process consisting of one or more steps. Each step
has input and output parameters defined by the inputs and outputs
fields. A workflow executes as described in execution model.
}
\section{Fields}{

\describe{
\item{\code{outputs}}{(WorkflowOutputParameterList) Defines the parameters
representing the output of the process. May be used to generate
and/or validate the output object. Inherited from Process}

\item{\code{steps}}{(WorkflowStepList) The individual steps that make up the
workflow. Steps are executed when all input data links are
fufilled. An implementation may choose to execute the steps in a
different order than listed and/or execute steps concurrently,
provided that dependencies between steps are met.}
}}
\section{Dependencies}{


Dependencies between parameters are expressed using the source
field on workflow step input parameters and workflow output
parameters.

The source field expresses the dependency of one parameter on
another such that when a value is associated with the parameter
specified by source, that value is propagated to the destination
parameter. When all data links inbound to a given step are
fufilled, the step is ready to execute.
}

\section{Extensions}{


ScatterFeatureRequirement and SubworkflowFeatureRequirement are
available as standard extensions to core workflow semantics.
}

