% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CWLAllClasses.R
\docType{class}
\name{WorkflowOutputParameter-class}
\alias{WorkflowOutputParameter}
\alias{WorkflowOutputParameter-class}
\title{WorkflowOutputParameter Class}
\description{
Describe an output parameter of a workflow. The parameter must be
connected to one or more parameters defined in the workflow that
will provide the value of the output parameter.
}
\section{Fields}{

\describe{
\item{\code{source}}{[character] Specifies one or more workflow parameters
that will provide this output value.}

\item{\code{linkMerge}}{[LinkMergeMethod] The method to use to merge
multiple inbound links into a single array. If not specified, the
default method is merge_nested:}
}}

