% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CWLAllClasses.R
\docType{class}
\name{WorkflowOutputParameterList-class}
\alias{WorkflowOutputParameterList}
\alias{WorkflowOutputParameterList-class}
\title{WorkflowOutputParameterList}
\description{
WorkflowOutputParameterList
}

