% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/07-Workflow-class.R
\name{WorkflowOutputParameterList}
\alias{WorkflowOutputParameterList}
\alias{WorkflowOutputParameterList-class}
\title{WorkflowOutputParameterList}
\usage{
WorkflowOutputParameterList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\description{
WorkflowOutputParameterList
}

