% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/70-Workflow-class.R
\docType{class}
\name{WorkflowStepInput-class}
\alias{WorkflowStepInput}
\alias{WorkflowStepInput-class}
\alias{WorkflowStepInputList}
\alias{WorkflowStepInputList-class}
\alias{WorkflowStepList}
\alias{WorkflowStepList-class}
\alias{WorkflowStepOutput}
\alias{WorkflowStepOutput-class}
\alias{WorkflowStepOutputList}
\alias{WorkflowStepOutputList-class}
\title{WorkflowStepInputList}
\usage{
WorkflowStepInputList(...)

WorkflowStepOutputList(...)

WorkflowStepList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\description{
WorkflowStepInputList

WorkflowStepOutputList

WorkflowStepList
}
\section{WorkflowStepInput Class}{

\describe{

The input of a workflow step connects an upstream parameter (from
the workflow inputs, or the outputs of other workflows steps) with
the input parameters of the underlying process.

#' If the sink parameter is an array, or named in a workflow scatter
operation, there may be multiple inbound data links listed in the
connect field. The values from the input links are merged depending
on the method specified in the linkMerge field. If not specified,
the default method is merge_nested:

\item{merge_nested}{ The input shall be an array consisting of
exactly one entry for each input link. If merge_nested is specified
with a single link, the value from the link is wrapped in a
single-item list.  }

\item{merge_flattened}{ 1) The source and sink parameters must be
compatible types, or the source type must be compatible with single
element from the "items" type of the destination array
parameter. 2) Source parameters which are arrays are concatenated;
source parameters which are single element types are appended as
single elements.  }

Fields:

\item{\code{id}}{ (character) A unique identifier for this workflow input
parameter.}

\item{\code{source}}{[character] Specifies one or more workflow parameters
that will provide input to the underlying process parameter.}

\item{\code{linkMerge}}{[LineMergeMethod] The method to use to merge
multiple inbound links into a single array. If not specified, the
default method is merge_nested:}

\item{\code{default}}{ [ANY] The default value for this parameter if there
is no source field.}
}
}

\section{WorkflowStepOutput Class}{

\describe{

Associate an output parameter of the underlying process with a
workflow parameter. The workflow parameter (given in the id field)
be may be used as a source to connect with input parameters of
other workflow steps, or with an output parameter of the process.

\item{\code{id}}{ (character) A unique identifier for this workflow output
parameter. This is the identifier to use in the source field of
WorkflowStepInput to connect the output value to downstream
parameters.}

}
}

