% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/07-Workflow-class.R
\docType{class}
\name{WorkflowStepInput-class}
\alias{WorkflowStepInput}
\alias{WorkflowStepInput-class}
\title{WorkflowStepInput Class}
\description{
The input of a workflow step connects an upstream parameter (from
the workflow inputs, or the outputs of other workflows steps) with
the input parameters of the underlying process.
}
\section{Fields}{

\describe{
\item{\code{id}}{(character) A unique identifier for this workflow input
parameter.}

\item{\code{source}}{[character] Specifies one or more workflow parameters
that will provide input to the underlying process parameter.}

\item{\code{linkMerge}}{[LineMergeMethod] The method to use to merge
multiple inbound links into a single array. If not specified, the
default method is merge_nested:}

\item{\code{default}}{[ANY] The default value for this parameter if there
is no source field.}
}}
\section{details}{


If the sink parameter is an array, or named in a workflow scatter
operation, there may be multiple inbound data links listed in the
connect field. The values from the input links are merged depending
on the method specified in the linkMerge field. If not specified,
the default method is merge_nested:

\itemize{

\item{merge_nested}{ The input shall be an array consisting of
exactly one entry for each input link. If merge_nested is specified
with a single link, the value from the link is wrapped in a
single-item list.  }

\item{merge_flattened}{ 1) The source and sink parameters must be
compatible types, or the source type must be compatible with single
element from the "items" type of the destination array
parameter. 2) Source parameters which are arrays are concatenated;
source parameters which are single element types are appended as
single elements.  } }
}

