% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/07-Workflow-class.R
\name{WorkflowStepInputList}
\alias{WorkflowStepInputList}
\alias{WorkflowStepInputList-class}
\title{WorkflowStepInputList}
\usage{
WorkflowStepInputList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\description{
WorkflowStepInputList
}

