% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/70-Workflow-class.R
\name{WorkflowStepList}
\alias{WorkflowStepList}
\alias{WorkflowStepList-class}
\title{WorkflowStepList}
\usage{
WorkflowStepList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\description{
WorkflowStepList
}

