% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/70-Workflow-class.R
\docType{class}
\name{WorkflowStepOutput-class}
\alias{WorkflowStepOutput}
\alias{WorkflowStepOutput-class}
\title{WorkflowStepOutput Class}
\description{
Associate an output parameter of the underlying process with a
workflow parameter. The workflow parameter (given in the id field)
be may be used as a source to connect with input parameters of
other workflow steps, or with an output parameter of the process.
}
\section{Fields}{

\describe{
\item{\code{id}}{(character) A unique identifier for this workflow output
parameter. This is the identifier to use in the source field of
WorkflowStepInput to connect the output value to downstream
parameters.}
}}

