% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/70-Workflow-class.R
\name{WorkflowStepOutputList}
\alias{WorkflowStepOutputList}
\alias{WorkflowStepOutputList-class}
\title{WorkflowStepOutputList}
\usage{
WorkflowStepOutputList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\description{
WorkflowStepOutputList
}

