% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/60-CommandLineTool-class.R
\name{CCBList}
\alias{CCBList}
\alias{characterORCommandLineBindingList-class}
\title{characterORCommandLineBindingList Class}
\usage{
CCBList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\value{
CCBList
}
\description{
characterORCommandLineBindingList Class
}
\examples{
CCBList("-o output.bam")
}

