% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/60-CommandLineTool-class.R
\docType{class}
\name{CommandInputParameter-class}
\alias{CommandInputParameter}
\alias{CommandInputParameter-class}
\title{CommandInputParameter Class}
\description{
An input parameter for a CommandLineTool.
}
\examples{
ipl <- InputParameterList(
    CommandInputParameter(id = "BAM", type = "File",
                          label = "input bam",
                          description = "input bam",
                          inputBinding = CommandLineBinding(
                              position = 1L
                          )),
    CommandInputParameter(id = "level", type = "Integer",
                          label = "Compression level",
                          description = "Compression level",
                          inputBinding = CommandLineBinding(
                              position = 2L,
                              prefix = "-l"
                          ))
)
}

