% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/40-Process-class.R
\name{ProcessRequirementList}
\alias{ProcessRequirementList}
\alias{ProcessRequirementList-class}
\title{ProcessRequirementList}
\usage{
ProcessRequirementList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\description{
ProcessRequirementList
}

